v0.1.0 "Laracasts From Scratch" video tutorial finished.
Current status of the blog:

Tought in the video series:
~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
OK Register User
OK Encrypt password
OK Logout
OK Login
OK Comments

OK Post Listing
OK Show Post
OK Sidebar, Archives links work and list is filtered
OK Sidebar, Tags links work and list is filtered.

Not told in the video series, so untouched( nonfunctional  )
~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
Pagination of Posts
Navigation Menu items not functional
About in sidebar is static
Elsewhere in sidebar is static
Footer is untouched , as taken from bootstrap
Add tags while creating a post
No admin panel

The intention of the video series was not creating a full functional site,
but teaching basics of Laravel as easy as possible.

Intention is fullfilled. Thanks Jeffrey.

Source of the video tutorial:
https://laracasts.com/series/laravel-from-scratch-2017

~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
Code at this point reflects the code told in the tutorial, as possible.
I made some changes to the code, but not much.
Summary of changes:
- Comments are working (after user auth added).
- A testing connection added for testing purposes.
- Some syntax errors fixed as needed.
~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~

This code will be my starting point. Will be changed as needed to
make a blog application.